% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briefSummary.R
\name{briefSummary}
\alias{briefSummary}
\title{Generating Summary Tables for levels of a categorical response.}
\usage{
briefSummary(x, response, maxsum = 7)
}
\arguments{
\item{x}{a data.frame to be summarized}

\item{response}{a character string containing the name of the categorical response variable.}

\item{maxsum}{a numerical value (default 7) determining the number of summary details generated.
For numerical predictors, this is passed to \code{summary()}, and for categorical variables,
this is the maximum number of levels shown in the output.}
}
\value{
\code{briefSummary} returns a list containing two objects, one a summary table
for the numerical variables (called \code{Numerical}) and the other a list of tables (called \code{Categorical}),
each table summarizing one categorical predictor.
}
\description{
This function returns summaries of numerical and categorical predictors
based on the levels of a categorical response variable.
}
