% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{calculate_file_hash}
\alias{calculate_file_hash}
\title{Calculate file hash for integrity checking}
\usage{
calculate_file_hash(file_path)
}
\arguments{
\item{file_path}{Path to the file}
}
\value{
MD5 hash of the file
}
\description{
Calculate file hash for integrity checking
}
