% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDS.R
\name{cleanDS}
\alias{cleanDS}
\title{Cleaning a Dataset}
\usage{
cleanDS(
  x,
  cleanDF = data.frame(Name = c("NACCID", "NACCVNUM", "NACCALZD", "NACCBMI", "NACCAGE",
    "HISPANIC"), CurrentVal = c("", "", "8,0,1", "", "", "0,1,9"), NewLevel = c("", "",
    "Normal,Dementia,Alzheimer", "", "", "Non-Hispanic, Hispanic, 9"), NA_Vals = c("",
    "", "", "\\"-4, 888.8\\"", "", "\\"9\\""), Rename = c("", "VisitNum", "AlzD", "BMI",
    "AGE", "Hispanic")),
  Rename = FALSE,
  addCols = NULL
)
}
\arguments{
\item{x}{the dataset to be cleaned; the default dataset is \code{UDS}}

\item{cleanDF}{a data.frame that includes instructions for cleaning x;
columns at least should include the following columns:
\itemize{
\item \code{Name}: Containing names of variables to be extracted,
\item \code{CurrentVal}, \code{NewLevel}: The current values and their replacement levels,
\item \code{NA_Vals}: The values or levels to be replaced with an NA.
\item \code{Rename}: Suggested new variable names, if needed,
See the examples for a CleanDF sample.
}}

\item{Rename}{a logical value indicating whether the variable names should
be replaced with values in the \code{Rename} column, if given.}

\item{addCols}{list of columns to include as they are.}
}
\value{
returns a selection of variables that have been cleaned with new
levels and NA values.
}
\description{
Cleaning a Dataset
}
\examples{
# Creating a sample cleanDF data frame:
 sampleCleanDF <- data.frame(
    Name=c('NACCID', 'NACCVNUM', 'NACCALZD', 'NACCBMI', 'NACCAGE', 'HISPANIC'),
    CurrentVal=c('','','8,0,1','','','0,1,9'),
    NewLevel=c('','','Normal,Dementia,Alzheimer','','','Non-Hispanic, Hispanic, 9'),
    NA_Vals=c('','','','"-4, 888.8"','','"9"'),
    Rename=c('','VisitNum', 'AlzD', 'BMI','AGE','Hispanic'))
# Extracting Data
cleanDS(x, cleanDF=sampleCleanDF, Rename=FALSE)

}
