% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{concat}
\alias{concat}
\title{Concatenating variables to create a new one}
\usage{
concat(x, formula)
}
\arguments{
\item{x}{a data frame}

\item{formula}{a response-less formula (~ followed by a \code{+}-separated list of
variables)}
}
\value{
A new variable by concatenating values of variables given as
predictors given in the \code{formula}.
}
\description{
A utility function for concatenates a \code{+}-separated list of variables.
}
\examples{
concat(data.frame(A=c('a', 'b', 'a', 'b'), B=c('x', 'y', 'y', 'x')), ~A+B)

}
