% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{cont_table}
\alias{cont_table}
\title{Contingency Table}
\usage{
cont_table(x, formula, useNA = "no")
}
\arguments{
\item{x}{A data frame}

\item{formula}{a formula of form \code{response(s) ~ predictor(s)} for constructing the
contingency table}
}
\value{
a contingency table which its rows and columns are identified by
the \code{formula}.
}
\description{
Creating a contingency table for based on a \code{response(s) ~ predictor(s)} formula.
}
\examples{
cont_table(data.frame(A=c('a', 'b', 'a', 'b'), B=c('x', 'y', 'y', 'x')), A~B))

}
\seealso{
\code{\link[=concat]{concat()}}
}
