% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{delete_var}
\alias{delete_var}
\title{Delete Variable}
\usage{
delete_var(obj, name)
}
\arguments{
\item{obj}{An object of class "sqlite_storage"}

\item{name}{Character string, name of the variable to delete
#$export}
}
\description{
Delete Variable
}
