% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_ds.R
\name{describe_ds}
\alias{describe_ds}
\title{Describe a Dataset}
\usage{
describe_ds(x, maxLevels = 7, var2row = FALSE)
}
\arguments{
\item{x}{a dataset}

\item{maxLevels}{the maximum number of levels for categorical variables (default is 7)}

\item{var2row}{if TRUE, each variable is summarized as a row; default is
FALSE, where each column represents the summary of a variable}
}
\value{
A list of at most two tables (Numerical and Categorical) containing
the summary information for variables of each type.
}
\description{
Describe a Dataset
}
