% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing.R
\name{fill_missing}
\alias{fill_missing}
\alias{fill_complement}
\title{Filling missing values
Replacing a missing value with prior values.}
\usage{
fill_missing(x, value = NULL, direction = "down")

fill_complement(x, y, fill_missing = "down")
}
\arguments{
\item{x}{a vector to be updated}

\item{value}{The value to be considered as missing; if missing, NA is used as
the default.}

\item{direction}{the direction where non-missing values are reused.}
}
\value{
The updated x vector where missing values are filled by existing
values.
}
\description{
Complements missing value of \code{x} with the values of \code{y}
}
