% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{get_provenance}
\alias{get_provenance}
\title{Get Variable Provenance}
\usage{
get_provenance(obj, name)
}
\arguments{
\item{obj}{An object of class "sqlite_storage"}

\item{name}{Character string, name of the variable}
}
\value{
Data frame with provenance information
#$export
}
\description{
Get Variable Provenance
}
