% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{initialize_db}
\alias{initialize_db}
\title{Initialize Database Connection}
\usage{
initialize_db(obj)
}
\arguments{
\item{obj}{An object of class "sqlite_storage"
#$export}
}
\description{
Initialize Database Connection
}
