% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{list_files}
\alias{list_files}
\title{List All Registered Files}
\usage{
list_files(obj)
}
\arguments{
\item{obj}{An object of class "sqlite_storage"}
}
\value{
A data frame with file information
#$export
}
\description{
List All Registered Files
}
