% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{list_vars}
\alias{list_vars}
\title{List All Variables with Provenance}
\usage{
list_vars(obj)
}
\arguments{
\item{obj}{An object of class "sqlite_storage"}
}
\value{
A data frame with variable information including provenance
#$export
}
\description{
List All Variables with Provenance
}
