% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{load_file_by_type}
\alias{load_file_by_type}
\title{Load file based on its type}
\usage{
load_file_by_type(file_path)
}
\arguments{
\item{file_path}{Path to the file}
}
\value{
Loaded R object
}
\description{
Load file based on its type
}
