% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{load_var}
\alias{load_var}
\title{Load Variable from Database}
\usage{
load_var(obj, name)
}
\arguments{
\item{obj}{An object of class "sqlite_storage"}

\item{name}{Character string, name of the variable to load}
}
\value{
The stored R object
#$export
}
\description{
Load Variable from Database
}
