% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{register_file}
\alias{register_file}
\title{Register a file in the database}
\usage{
register_file(obj, file_path)
}
\arguments{
\item{obj}{An object of class "sqlite_storage"}

\item{file_path}{Path to the file to register
#$export}
}
\description{
Register a file in the database
}
