% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{save_file_var}
\alias{save_file_var}
\title{Save Variable from File with Processing Steps}
\usage{
save_file_var(obj, name, file_path, steps = NULL, processor = NULL)
}
\arguments{
\item{obj}{An object of class "sqlite_storage"}

\item{name}{Character string, name of the variable}

\item{file_path}{Path to the file to load}

\item{steps}{Optional, character vector describing processing steps}

\item{processor}{Optional, custom function to process the file. If NULL, uses default processors.
#$export}
}
\description{
Save Variable from File with Processing Steps
}
