% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{save_file_var.sqlite_storage}
\alias{save_file_var.sqlite_storage}
\title{#$export}
\usage{
\method{save_file_var}{sqlite_storage}(obj, name, file_path, steps = NULL, processor = NULL)
}
\description{
#$export
}
