% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{save_var}
\alias{save_var}
\title{Save Variable to Database with Provenance}
\usage{
save_var(obj, name, value, source_file = NULL, steps = NULL)
}
\arguments{
\item{obj}{An object of class "sqlite_storage"}

\item{name}{Character string, name of the variable}

\item{value}{Any R object to be stored}

\item{source_file}{Optional, path to source file used to create this variable}

\item{steps}{Optional, character vector describing the steps taken to create this variable
#$export}
}
\description{
Save Variable to Database with Provenance
}
