% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{save_variable_steps}
\alias{save_variable_steps}
\title{Save detailed steps for a variable}
\usage{
save_variable_steps(obj, var_name, steps)
}
\arguments{
\item{obj}{An object of class "sqlite_storage"}

\item{var_name}{Variable name}

\item{steps}{Character vector of steps}
}
\description{
Save detailed steps for a variable
}
