% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSummary.R
\name{tableSummary}
\alias{tableSummary}
\alias{col_total.table}
\alias{row_total.table}
\title{Summarizing Functions for \code{table} Object}
\usage{
col_total.table(x)

row_total.table(x)
}
\arguments{
\item{x}{a table object.}
}
\value{
\code{col_total} returns a new table object with an additional column,
named \code{Total} or \code{Total.Col}

\code{row_total} returns a new table object with an additional row,
named \code{Total} or \code{Total.Row}
}
\description{
Utility functions for adding Totals

\code{col_total.table} adds a column summary to a table.

\code{row_total.table} Adds a row summary to a table.
}
