% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Utility functions for internal use}
\alias{Utility functions for internal use}
\alias{calculate_file_hash}
\alias{concat}
\title{Internal helper functions}
\usage{
calculate_file_hash(file_path)

concat(x, formula)
}
\arguments{
\item{file_path}{Path to the file}

\item{x}{a data frame where variables are taken from.}

\item{formula}{a response-less formula (~ followed by a \code{+}-separated list of
variables)}
}
\value{
MD5 hash of the file

A new variable by concatenating values of variables given by \code{formula}.
}
\description{
Calculate file hash for integrity checking

A utility function that creates a new variable by joining the values of a given
list of variables.
}
\examples{
du4tnr:::concat(data.frame(A=c('a', 'b', 'a', 'b'), B=c('x', 'y', 'y', 'x')), ~A+B)

}
\keyword{internal}
