% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_storage.R
\name{var_exists}
\alias{var_exists}
\title{Check if Variable Exists}
\usage{
var_exists(obj, name)
}
\arguments{
\item{obj}{An object of class "sqlite_storage"}

\item{name}{Character string, name of the variable}
}
\value{
Logical, TRUE if variable exists
#$export
}
\description{
Check if Variable Exists
}
