% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varClean.R
\name{varClean}
\alias{varClean}
\alias{setNAs}
\title{Variable cleaning functions}
\usage{
setNAs(x, naVals = NA)
}
\arguments{
\item{x}{a vector to be updated with some new \code{NA}'s.}

\item{naVals}{a vector of values that should be considered as \code{NA}.
The default value is \code{NA} (no new NA's will be introduced).}
}
\value{
a modified version of \code{x} where values in \code{naVals} are replaced with \code{NA}'s.
}
\description{
A set of functions for cleaning variables in a dataset.

\code{setNAs} is used to set a list of given values as \code{NA}.
}
