% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_ds.R
\name{which_ds}
\alias{which_ds}
\alias{whichds}
\title{Searching a column name in datasets (data frames)}
\usage{
whichds(columnName = NULL, ...)

which_ds(columnName = NULL, ...)
}
\arguments{
\item{columnName}{a character vector of name of variables of interest.}

\item{...}{a comma separated list of datasets}

\item{dss}{is a list of all datasets where variable names will be searched for.}
}
\value{
is a named, logical array that determines which dataset includes
variables listed in \code{columnName}.
}
\description{
A helper function for looking up one variable name in a list of datasets.

Looking up the name of several variables in a list of datasets.
}
