% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvalid.R
\name{xvalid}
\alias{xvalid}
\alias{mcrBin}
\alias{misMul}
\alias{misMulOH}
\alias{mse}
\alias{cvGlm}
\title{K-Fold Cross Validation and Accuracy Functions}
\usage{
mcrBin(yhat, y, cutoff = 0.5, FUN = mean)

misMul(yhat, y, FUN = mean)

misMulOH(yhat, y, FUN = mean)

mse(yhat, y)

cvGlm(
  formula,
  data,
  K = 10,
  cost = accBin,
  method = glm,
  predType = "response",
  FUN = mean,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{yhat}{the predicted value}

\item{y}{the actual value}

\item{cutoff}{the value used to convert a probability into a 0/1 output.
The default cutoff value is 0.5.}

\item{FUN}{the function(s) used to calculate the summary of predictions. The
default value is \code{mean}.}

\item{formula}{the model formula in form of y~f(X), where y is
the response and f(x) is a function of predictors in a vector X.}

\item{data}{the dataset containing the response y and predictors X.}

\item{K}{the number of folds to be used in the cross validation.}

\item{cost}{the function that measures the accuracy or loss. The default
value is \code{acc} function.}

\item{method}{the method that fits the data to the formula and returns
a model object. The default value is \code{glm}.}

\item{predType}{is the type of prediction that is needed for the \code{cost} model
to work properly. The default value is 'response'.}

\item{...}{other parameters; all are passed to \code{method} function.}
}
\value{
\code{mcrBin} returns the rate of mis-predictions at \code{cutoff} level.

\code{misMul} returns the mis-prediction rate in a multi-level categorical variable.

\code{mse} returns the mean squared error (MSE) value for the predictions.

\code{cvGlm} returns an array of \code{K} calculated \code{cost} values computed for the train/test
pairs generated from the data.
}
\description{
A set of functions to help performing cross-validation using different
\code{method} (default=glm).

\code{mcrBin}: computing mis-classification rate for a probability prediction of
a bi-level variable.

\code{misMul}: Mis-prediction for a multi-level factor response

\code{misMulOH}: returns the mis-predictions made for multiple-level response in one-hot format

\code{mse}: computes the mean squared deviation of the predictor from the actual value.

\code{cvGlm}: performing a K-fold cross-validation (using GLM method by default) on
\code{data} for a model defined by \code{formula}.
}
