% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_ds.R
\name{describe_ds}
\alias{describe_ds}
\title{Describe a Dataset}
\usage{
describe_ds(x, maxsum = 7, long = FALSE)
}
\arguments{
\item{x}{a dataset}

\item{maxsum}{the maximum number of summary items to supply}
}
\value{
A list of at most two tables (Numerical and Categorical) containing
the summary information for variables of each type.
}
\description{
Describe a Dataset
}
