% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NACCdata-package.R
\docType{package}
\name{du4tnr-package}
\alias{du4tnr}
\alias{du4tnr-package}
\title{du4tnr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
