% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NACCdata-package.R
\docType{package}
\name{du4tnr-package}
\alias{du4tnr}
\alias{du4tnr-package}
\title{du4tnr: Data Utility for Transparency and Reproducibility (DU4TnR)}
\description{
A set of tools, inclduing a SQLite-based variable storage object, for increased transparency and reproducibility in collaborative data analysis.
}
\author{
\strong{Maintainer}: Mansour Abdoli \email{mabdoli@fullerton.edu} (\href{https://orcid.org/0009-0001-6722-0134}{ORCID})

}
\keyword{internal}
