% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_ds.R
\name{which_ds}
\alias{which_ds}
\alias{whichds}
\title{Find the dataset containing given column names}
\usage{
whichds(columnName = NULL, dss)
}
\arguments{
\item{columnName}{a vector of variable names to be looked up is different datasets.}

\item{dss}{is a list of all datasets where variable names will be searched for.}
}
\value{
is a named, logical array that determines which dataset includes
variables listed in \code{columnName}.
}
\description{
A helper function for looking up one variable name in a list of datasets.
}
